requirements.txt
streamlit
biopython
pandas
numpy
matplotlib
nest_asyncio
